backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
flake8==7.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
sqlparse==0.5.3
whitenoise==6.9.0
